==> projects/kaggle-airbnb/notebooks/0_eda.ipynb <==
## Notebook to generate Pandas Profiling reports, for better Exploratory Data Analysis

#### To install pandas profiling package, run the following command in Jupyter cell:  
! pip install pandas-profiling[notebook]

import pandas as pd
from pandas_profiling import ProfileReport
from datetime import datetime

pd.options.display.float_format = "{:.2f}".format
%load_ext autotime

### 1. Creating profiling report for age_gender_bkts file

df = pd.read_csv('../data/original/age_gender_bkts.csv')
df.shape

df.head()

df[df.country_destination == 'AU'].population_in_thousands.sum()

df[(df.country_destination == 'AU') & (df.gender == 'female')]

profile = ProfileReport(df, title="Age Gender Buckets Profiling Report", explorative=True)
profile.to_file('../reports/age_gender_bkts.html')

### 2. Viewing countries file

df = pd.read_csv('../data/original/countries.csv')
df.shape

df

### 3. Creating profiling report for entire sessions file failed, creating for its subsample

df = pd.read_csv('../data/original/sessions.csv')
df.shape

df.head()

df.describe()

profile = ProfileReport(df.sample(100_000).reset_index(drop=True), title="Users Profiling Report", explorative=True)
profile.to_file('../reports/sessions.html')

### 4. Creating profiling report for train_users_2 file

df = pd.read_csv('../data/original/train_users_2.csv')
df.shape

df.head()

df.date_account_created = df.date_account_created.apply(lambda x: datetime.strptime(x, '%Y-%m-%d'))

df.timestamp_first_active = df.timestamp_first_active.apply(lambda x: datetime.strptime(str(x), '%Y%m%d%H%M%S'))

df.date_first_booking = df.date_first_booking.apply(lambda x: datetime.strptime(x, '%Y-%m-%d') if pd.notna(x) else None)

df.head()

profile = ProfileReport(df, title="Users Profiling Report", explorative=True)
profile.to_file('../reports/train_users_2.html')

df[df.age > 1000].country_destination.value_counts()

df[df.date_first_booking.isna()].country_destination.value_counts()

\[ DCG_k=\sum_{i=1}^k\frac{2^{rel_i}-1}{\log_2{\lefti+1\righti+1\right}}, \]
\[ nDCG_k=\frac{DCG_k}{IDCG_k}, \]

### 5. Creating profiling report for train_users_2 file

df = pd.read_csv('../data/original/test_users.csv')
df.shape

df.head()

df.date_account_created = df.date_account_created.apply(lambda x: datetime.strptime(x, '%Y-%m-%d'))

df.timestamp_first_active = df.timestamp_first_active.apply(lambda x: datetime.strptime(str(x), '%Y%m%d%H%M%S'))

df.date_first_booking = df.date_first_booking.apply(lambda x: datetime.strptime(x, '%Y-%m-%d') if pd.notna(x) else None)

df.head()

profile = ProfileReport(df, title="Test Users Profiling Report", explorative=True)
profile.to_file('../reports/test_users.html')
