==> projects/kaggle-airbnb/notebooks/5_features_filter.ipynb <==
### Filtering features from Low Variance, High Correlation and Empty Values Columns

import sys
sys.path.append('../src/')
import filter_features as ff
import pandas as pd

df = pd.read_parquet('../data/processed/train_features.parquet')
df.shape

empty_columns = ff.get_empty_columns(df, threshold=0.999)
len(empty_columns)

### 1. Getting and loading Feature Importance file

model_path = '/home/jovyan/projects/kaggle-airbnb/models/model_2021_09_06_14_22_11.cbm'
x_pickle_path = '/home/jovyan/projects/kaggle-airbnb/data/processed/x_test.pickle'
y_pickle_path = '/home/jovyan/projects/kaggle-airbnb/data/processed/y_test.pickle'
cat_features = [
    'gender',
    'signup_method',
    'signup_flow',
    'language',
    'affiliate_channel',
    'affiliate_provider',
    'first_affiliate_tracked',
    'signup_app',
    'first_device_type',
    'first_browser',
    'dow_registered',
    'hr_registered',
    'age_group',
    'dow_registered',
    'day_registered',
    'month_registered',
    'year_registered',
]
save_to = '/home/jovyan/projects/kaggle-airbnb/data/processed/features_importance.csv'

ff.generate_feature_importance_file(model_path, x_pickle_path, y_pickle_path, cat_features, save_to)

fi = ff.get_normalized_feature_weights(save_to)
fi = [{'feature_name': k, 'weight': v} for k, v in fi.items()]
fi = pd.DataFrame(fi)
fi['w_sum'] = fi.weight.cumsum()
fi.shape

fi.head(10)

fi[fi.feature_name == 'gender']

### 1.1 If we remove empty_cols columns we would lose approximately 1.4% of feature importance

fi[fi.feature_name.isin(empty_columns)].weight.sum()

### 2. Generating Low variance columns

low_variance_columns = ff.get_low_variance_columns(df, threshold=0.0001)
len(low_variance_columns)

fi[fi.feature_name.isin(low_variance_columns)].weight.sum()

remove_columns = set(empty_columns).union(set(low_variance_columns))
len(remove_columns)

'country_destination' in remove_columns

### 3. Generating highly correlated columns set

fi_weights = ff.get_normalized_feature_weights(save_to)
len(fi_weights), fi_weights['age']

df.drop(remove_columns, axis=1, inplace=True)
df.shape

cdf = ff.get_correlated_features_pairs(df, fi_weights, threshold=0.95)
cdf.shape

cdf.head()

ccs = ff.get_connected_components(cdf)
len(ccs)

ccs[-5]

corr = abs(df.corr())

ix = -5
corr[ccs[ix]][corr.index.isin(ccs[ix]) | corr.index.isin(ccs[ix])]

fi[fi.feature_name.isin(ccs[ix])].head()

# len(highly_correlated_features)

highly_correlated_features = ff.get_highly_correlated_removal_candidates(ccs, fi_weights, verbose=1)

remove_columns = remove_columns.union(set(highly_correlated_features))
len(remove_columns)

'country_destination' in remove_columns

### 4. Dropping filtered columns and saving data. Repeating same process for test set

remove_columns = list(set(remove_columns).intersection(set(df)))
df.drop(remove_columns, axis=1, inplace=True)
df.shape

test_df = pd.read_parquet('../data/processed/test_features.parquet')
test_df.shape

test_df.drop(remove_columns, axis=1, inplace=True)
test_df.shape

df.to_parquet('../data/processed/train_features_uncorr.parquet')

test_df.to_parquet('../data/processed/test_features_uncorr.parquet')

fi[fi.feature_name.str.startswith('ai_')].head(20)

users = pd.read_parquet('../data/processed/users.parquet')
users.shape

sessions = pd.read_parquet('../data/processed/sessions.parquet')
sessions.shape

df = pd.read_parquet('../data/processed/train_features.parquet')
df.shape

df.country_destination.value_counts(normalize=True)

df.country_destination.value_counts(dropna=False)

(NDF, US, other) - (Universum - (NDF, US, other))

df.head()

'action_info' in list

df.country_destination.value_counts()

other_countries_ai = df[~df.country_destination.isin({'NDF', 'US', 'other'})].action_info.tolist()
