==> projects/kaggle-airbnb/notebooks/6_undersampling.ipynb <==
### An attempt to balance classes in training dataset

import sys
sys.path.append('../src/')
import filter_features as ff
import pandas as pd

df = pd.read_parquet('../data/processed/train_features.parquet')
df.shape

df['nan_counts'] = df.isnull().sum(axis=1)

df[['country_destination', 'nan_counts']].head()

df.country_destination.value_counts()

df = df.sort_values(by=['country_destination', 'nan_counts'], ascending=[True, True])
df.reset_index(drop=True, inplace=True)

df.head()

tmp = df.groupby('country_destination', as_index=False).head(10000)
tmp.shape

tmp.country_destination.value_counts()

df.to_parquet('../data/processed/train_features_undersample.parquet')
df.shape
