==> projects/kaggle-airbnb/notebooks/9_generate_submission.ipynb <==
## Generatig predicitons for the submission dataset

from catboost import CatBoostClassifier, Pool
import pandas as pd
from collections import Counter
from tqdm.notebook import tqdm
import numpy as np
import glob

### 0. Loading and preparing data

# df = pd.read_parquet('../data/processed/test_features.parquet')
# df = pd.read_parquet('../data/processed/test_features_uncorr.parquet')
df = pd.read_parquet('../data/processed/test_features.parquet')
df.shape

df['nan_counts'] = df.isnull().sum(axis=1)

if 'train_flag' in df:
    df.drop('train_flag', inplace=True, axis=1)
    df.reset_index(drop=True, inplace=True)
    df.shape

df.head()

df.drop('country_destination', axis=1, inplace=True)

x = df.drop('user_id', axis=1)

cat_features = [
    'gender',
    'signup_method',
    'signup_flow',
    'language',
    'affiliate_channel',
    'affiliate_provider',
    'first_affiliate_tracked',
    'signup_app',
    'first_device_type',
    'first_browser',
    'dow_registered',
    'hr_registered',
    'age_group',
    'dow_registered',
    'day_registered',
    'month_registered',
    'year_registered',
]

cat_features_remained = list(set(cat_features).intersection(set(df)))
cat_features_removed = set(cat_features) - set(df)
len(cat_features), len(cat_features_remained), len(cat_features_removed)                                      

for col in cat_features_remained:
    x[col].fillna('', inplace=True)
    x[col] = x[col].astype('category')

### 2. Loading model

def find_latest_model(path='../models/', ext='cbm'):
    files = glob.glob(f"{path}*.{ext}")
    files = [file for file in files if len(file) > 25]
    files = sorted(files)
    return files[-1]

path = find_latest_model()
path

model = CatBoostClassifier()
model.load_model(path)

### 3.3 Predicting 5 Countries of Destination per each user

classes = list(model.classes_)
classes

preds = model.predict_proba(x)
preds = preds.tolist()

preds_df = pd.DataFrame({'preds': preds})
preds_df['amax'] = preds_df.preds.apply(lambda x: max(x))

preds_df.head()

def find_max_n_indeces(x, n=1):
    n_indeces = sorted(x, reverse=True)[:n]
    n_indeces = [x.index(el) for el in n_indeces]
    return n_indeces


def get_best_n_result(x, n=5):
    n_indeces = find_max_n_indeces(x, n=n)
    best_n = [classes[ix] for ix in n_indeces]
    return best_n

find_max_n_indeces(preds_df.loc[0].preds, n=5), get_best_n_result(preds_df.loc[0].preds, n=5)

find_max_n_indeces(preds_df.loc[0].preds, n=5), get_best_n_result(preds_df.loc[0].preds, n=5)

preds_df['best5'] = preds_df.preds.apply(lambda x: get_best_n_result(x, n=5))

preds_df.head()

preds_df.head()

preds_df['id'] = df['user_id']

preds_df.head()

submission = preds_df[['id', 'best5']].explode('best5')
submission.shape

submission.head()

### 4 Saving NDCG aware submission

submission.columns = ['id', 'country']

path[-24:-4]

submission.to_csv(f'../data/results/submission{path[-24:-4]}.csv', index=False)

def get_sum_of_top_n(x, n=1):
    return sum(sorted(x, reverse=True)[:n])

preds_df['sum5'] = preds_df.preds.apply(lambda x: get_sum_of_top_n(x, 5))
preds_df['sum4'] = preds_df.preds.apply(lambda x: get_sum_of_top_n(x, 4))

preds_df.head()

s = preds_df[['id', 'preds']].explode('preds').copy(deep=True)
s.shape

s.preds = s.preds.astype(float)

classes

s['country'] = classes * int(len(s) / len(classes))

s.head()

type(s.loc[0].preds)

s = s.sort_values(['id', 'preds'], ascending=[True, False])
s.reset_index(drop=True, inplace=True)

ss = s[['id', 'preds']].groupby('id', as_index=False).preds.cumsum()

s['preds_sum'] = ss

s.head(15)

best = s.groupby('id', as_index=False).head(4)
best.shape

best.head()

result = s[s.preds_sum < 0.90]
result.shape, s.shape, result.id.nunique(), s.id.nunique()

result = pd.concat([best, result])
result = result.drop_duplicates(['id', 'country'])
result.shape, s.shape, result.id.nunique(), s.id.nunique()

result[['id', 'country']].to_csv('../data/results/submission_cumsum_head3_90.csv', index=False)
result.shape

'gender' in df

df.shape

users = pd.get_dummies(df[['gender', 'user_id', 'country_destination']], columns=['gender'])

users.head()
